+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- `itemBorderRadius` parameter to `TypeAheadField` and `CupertinoTypeAheadField` (#584)
+
+### Fixed
+- Mouse cursor doesn't change to pointer-hand when ListTile is returned in itemBuilder (#584)
+
 ## 5.2.0 - 2024-02-08
 ### Added
 - force refreshing suggestions with `SuggestionsController.refresh`